add checkpermissions to menu

+++ backend/db/controllers/menu.controller.js
@@ -2,50 +2,57 @@ const { Menu } = require('../models');
 const patchValidation = { runValidators: true };
 
 
-exports.postMenu = function(req, res) {
-    let startDate = req.body.startDate;
-    let endDate = req.body.endDate;
-    let dateStart = new Date(startDate);
-    let dateEnd = new Date(endDate);
-    let meals = req.body.meals;
-    let menuName = req.body.menuName;
+exports.postMenu = function (req, res) {
+    if (checkPermissions(req)) {
 
-    let newMenu = new Menu({
-        startDate,
-        endDate,
-        meals,
-        menuName
-    });
-    newMenu.save()
-        .then((menuDoc) => {
-            res.send(menuDoc);
+
+        let startDate = req.body.startDate;
+        let endDate = req.body.endDate;
+        let dateStart = new Date(startDate);
+        let dateEnd = new Date(endDate);
+        let meals = req.body.meals;
+        let menuName = req.body.menuName;
+
+        let newMenu = new Menu({
+            startDate,
+            endDate,
+            meals,
+            menuName
+        });
+        newMenu.save()
+            .then((menuDoc) => {
+                res.send(menuDoc);
+            }).catch((e) => {
+                res.send(e);
+            });
+    }
+}
+
+exports.deleteMenu = function (req, res) {
+    if (checkPermissions(req)) {
+        Menu.findOneAndRemove({
+            _id: req.params.id
+        }).then((removedMenuDoc) => {
+            res.send(removedMenuDoc);
         }).catch((e) => {
             res.send(e);
         });
+    }
 }
 
-exports.deleteMenu = function(req, res) {
-    Menu.findOneAndRemove({
-        _id: req.params.id
-    }).then((removedMenuDoc) => {
-        res.send(removedMenuDoc);
-    }).catch((e) => {
-        res.send(e);
-    });
+exports.patchMenu = function (req, res) {
+    if (checkPermissions(req)) {
+        Menu.findOneAndUpdate({ _id: req.params.id }, {
+            $set: req.body
+        }, patchValidation).then(() => {
+            res.send({ message: 'Completed successfully' });
+        }).catch((e) => {
+            res.send(e);
+        });
+    }
 }
 
-exports.patchMenu = function(req, res) {
-    // Lägg till datum-validering här också
-    Menu.findOneAndUpdate({ _id: req.params.id }, {
-        $set: req.body
-    }, patchValidation).then(() => {
-        res.send({ message: 'Completed successfully' });
-    }).catch((e) => {
-        res.send(e);
-    });
-}
-
-exports.getMenus = function(req, res) {
+exports.getMenus = function (req, res) {
     Menu.find().then((menus) => {
         res.send(menus);
     }).catch((e) => {
@@ -53,7 +60,7 @@ exports.getMenus = function(req, res) {
     });
 }
 
-exports.getMenuNames = function(req, res) {
+exports.getMenuNames = function (req, res) {
     Menu.find().select({ 'menuName': 1 }).then((menus) => {
         res.send(menus);
     }).catch((e) => {
@@ -61,7 +68,7 @@ exports.getMenuNames = function(req, res) {
     });
 }
 
-exports.getMenuWithId = function(req, res) {
+exports.getMenuWithId = function (req, res) {
     Menu.findOne({ _id: req.params.id }).then((menu) => {
         res.send(menu);
     }).catch((e) => {
@@ -69,7 +76,7 @@ exports.getMenuWithId = function(req, res) {
     });
 }
 
-exports.getMenuNameWithId = function(req, res) {
+exports.getMenuNameWithId = function (req, res) {
     Menu.findOne({ _id: req.params.id }).select({ 'menuName': 1 }).then((menus) => {
         res.send(menus);
     }).catch((e) => {
@@ -77,39 +84,43 @@ exports.getMenuNameWithId = function(req, res) {
     });
 }
 
-exports.deleteMeal = function(req, res) {
-    Menu.findOneAndUpdate({
-        _id: req.params.id
-    }, {
-        $pull: {
-            'meals': {
-                _id: req.params.mealId
+exports.deleteMeal = function (req, res) {
+    if (checkPermissions(req)) {
+        Menu.findOneAndUpdate({
+            _id: req.params.id
+        }, {
+            $pull: {
+                'meals': {
+                    _id: req.params.mealId
+                }
             }
-        }
-    }).then(() => {
-        res.send({ message: 'Completed successfully' });
-    }).catch((e) => {
-        res.send(e);
-    });
-
+        }).then(() => {
+            res.send({ message: 'Completed successfully' });
+        }).catch((e) => {
+            res.send(e);
+        });
+    }
 }
 
-exports.postMeal = function(req, res) {
-    Menu.findOneAndUpdate({
-        _id: req.params.id,
-    }, {
-        $push: {
-            'meals': req.body
-        }
-    }).then(() => {
-        res.send({ message: 'Completed successfully' });
-    }).catch((e) => {
-        res.send(e);
-    });
+exports.postMeal = function (req, res) {
+    if (checkPermissions(req)) {
+        Menu.findOneAndUpdate({
+            _id: req.params.id,
+        }, {
+            $push: {
+                'meals': req.body
+            }
+        }).then(() => {
+            res.send({ message: 'Completed successfully' });
+        }).catch((e) => {
+            res.send(e);
+        });
+    }
 }
 
-exports.patchMeal = function(req, res) {
-    Menu.findOneAndUpdate({
+exports.patchMeal = function (req, res) {
+    if (checkPermissions(req)) {
+        Menu.findOneAndUpdate({
             _id: req.params.id,
             "meals._id": req.params.mealId
         }, {
@@ -122,16 +133,19 @@ exports.patchMeal = function(req, res) {
         .catch((err) => {
             res.send(err);
         });
+    }
 }
 
 function checkPermissions(req) {
-    if(req.headers.usermail) {
-        return User.findOne({"email": req.headers.usermail}, function (err, myUser){
-            if(err){
+    if (req.headers.usermail) {
+        return User.findOne({ "email": req.headers.usermail }, function (err, myUser) {
+            if (err) {
                 return false;
-            } else if (myUser.permissions.some((perm) => perm === 'admin')){
+            } else if (myUser.permissions.some((perm) => perm === 'admin')) {
                 return true;
-            }
+            } else if(myUser.menuIds.some((id) => id === req.params.id)) {
+                return true
+            } else return false;
         })
     } else {
         return false;
